feat: remodelando jeito de fazer as migrations para poder implementar de outra forma(agora funciona)
