+++ database/models/RoomHasSchedule.ts
